requirements.txt
numpy
PyPDF2
python-docx
streamlit
transformers
langchain
langchain-community
tf-keras
sentencepiece